+++ pkg/index/reflect.go
@@ -1,0 +1,25 @@
+package index
+
+import (
+	"errors"
+	"reflect"
+)
+
+func getType(v interface{}) (reflect.Value, error) {
+	rv := reflect.ValueOf(v)
+	for rv.Kind() == reflect.Ptr || rv.Kind() == reflect.Interface {
+		rv = rv.Elem()
+	}
+	if !rv.IsValid() {
+		return reflect.Value{}, errors.New("failed to read value via reflection")
+	}
+
+	return rv, nil
+}
+
+func valueOf(v interface{}, field string) string {
+	r := reflect.ValueOf(v)
+	f := reflect.Indirect(r).FieldByName(field)
+
+	return f.String()
+}
